Add parameter for additional endpoints
This will allow passing in of different endpoints that might not be part
of a build (such as wanting to point to a remote API server instead of a
locally built one. Might be useful to not need to include (and
build/run) everything for larger projects... or for UAT/Integration
tests.